7-managed-rag/requirements.txt
langchain-community
langchain-openai
langchain-core
langgraph
azure-search-documents
python-dotenv
